edit category activity, adapter (set id type as lonh without int )

+++ app/src/main/java/artemshumidub/ru/sebbianews/ui/adapter/CategoryRVAdapter.java
@@ -63,7 +63,7 @@ public class CategoryRVAdapter extends RecyclerView.Adapter<CategoryRVAdapter.Ho
     }
 
     public interface OnItemListener{
-        void onItemClick(int id);
+        void onItemClick(long id);
     }
 
 }
